requirements.txt
flask-githubapp==0.2.0
GitHub-Flask==3.2.0
flask_sqlalchemy==2.4.3
flask-wtf==0.41.3
email-validator==1.1.1